feat(server): adding WithSession option
